Refactor visualization functions: migrate to vut.visualization module and mark deprecated
- Updated `plot_palette`, `plot_image`, `plot_images`, `plot_feature`, `plot_features`, `plot_scatter`, `plot_metrics`, `plot_action_segmentation`, and `plot_roc_curve` functions to use the new `vut.visualization` module.
- Added deprecation warnings for each function, indicating they will be removed in future versions.
- Removed direct imports of matplotlib and numpy, relying on the new visualization module for plotting.
- Ensured all function signatures and parameters remain consistent with previous implementations.
